# docs/swagger.yaml
basePath: /
definitions:
...
  managers.Response:
    properties:
      code:
        type: integer
      message:
        type: string
      success:
        type: boolean
    type: object
  managers.ResponseID:
    properties:
      code:
        type: integer
      id:
        type: integer
      message:
        type: string
      success:
        type: boolean
    type: object
...
  six910-database-interface.Category:
    properties:
      desc:
        type: string
      id:
        type: integer
      image:
        type: string
      name:
        type: string
      parentCategoryId:
        type: integer
      storeId:
        type: integer
      thumbnail:
        type: string
    type: object
...
host: localhost:3002
info:
  contact:
    name: API Support
    url: http://www.ulboralabs.com/contact/form
  description: This is the Six910 (six nine ten) API
  license:
    name: GPL-3.0
    url: https://github.com/Ulbora/Six910/blob/master/LICENSE
  termsOfService: https://github.com/Ulbora/Six910/blob/master/tos.html
  title: Six910 API
  version: "1.0"
paths:
...
  /rs/category/add:
    post:
      consumes:
      - application/json
      description: Adds a new category to a store
      parameters:
      - description: category
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/six910-database-interface.Category'
      - description: apiKey required for non OAuth2 stores only
        in: header
        name: apiKey
        type: string
      - description: store name
        in: header
        name: storeName
        required: true
        type: string
      - description: store localDomain
        in: header
        name: localDomain
        required: true
        type: string
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: OAuth2 client ID only for OAuth2 stores
        in: header
        name: clientId
        type: string
      - description: User ID only for OAuth2 stores
        in: header
        name: userId
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/managers.ResponseID'
      summary: Add a new category
      tags:
      - Category
  /rs/category/delete/{id}/{storeId}:
    delete:
      consumes:
      - application/json
      description: Delete a category from the store
      parameters:
      - description: category id
        in: path
        name: id
        required: true
        type: string
      - description: store storeId
        in: path
        name: storeId
        required: true
        type: string
      - description: apiKey required for non OAuth2 stores only
        in: header
        name: apiKey
        type: string
      - description: store name
        in: header
        name: storeName
        required: true
        type: string
      - description: store localDomain
        in: header
        name: localDomain
        required: true
        type: string
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: OAuth2 client ID only for OAuth2 stores
        in: header
        name: clientId
        type: string
      - description: User ID only for OAuth2 stores
        in: header
        name: userId
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/managers.Response'
      summary: Delete a category
      tags:
      - Category
  /rs/category/get/id/{id}/{storeId}:
    get:
      consumes:
      - application/json
      description: Get details of a category
      parameters:
      - description: category id
        in: path
        name: id
        required: true
        type: string
      - description: store storeId
        in: path
        name: storeId
        required: true
        type: string
      - description: apiKey required for non OAuth2 stores only
        in: header
        name: apiKey
        type: string
      - description: store name
        in: header
        name: storeName
        required: true
        type: string
      - description: store localDomain
        in: header
        name: localDomain
        required: true
        type: string
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: OAuth2 client ID only for OAuth2 stores
        in: header
        name: clientId
        type: string
      - description: User ID only for OAuth2 stores
        in: header
        name: userId
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/six910-database-interface.Category'
      summary: Get details of a category by id
      tags:
      - Category
  /rs/category/get/list/{storeId}:
    get:
      consumes:
      - application/json
      description: Get list of categories for a store
      parameters:
      - description: store storeId
        in: path
        name: storeId
        required: true
        type: string
      - description: apiKey required for non OAuth2 stores only
        in: header
        name: apiKey
        type: string
      - description: store name
        in: header
        name: storeName
        required: true
        type: string
      - description: store localDomain
        in: header
        name: localDomain
        required: true
        type: string
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: OAuth2 client ID only for OAuth2 stores
        in: header
        name: clientId
        type: string
      - description: User ID only for OAuth2 stores
        in: header
        name: userId
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/six910-database-interface.Category'
            type: array
      summary: Get list of categories
      tags:
      - Category
  /rs/category/get/sub/list/{catId}:
    get:
      consumes:
      - application/json
      description: Get list of sub categories for a store
      parameters:
      - description: category id
        in: path
        name: catId
        required: true
        type: string
      - description: apiKey required for non OAuth2 stores only
        in: header
        name: apiKey
        type: string
      - description: store name
        in: header
        name: storeName
        required: true
        type: string
      - description: store localDomain
        in: header
        name: localDomain
        required: true
        type: string
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: OAuth2 client ID only for OAuth2 stores
        in: header
        name: clientId
        type: string
      - description: User ID only for OAuth2 stores
        in: header
        name: userId
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/six910-database-interface.Category'
            type: array
      summary: Get list of sub categories
      tags:
      - Category
  /rs/category/update:
    put:
      consumes:
      - application/json
      description: Update category data
      parameters:
      - description: category
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/six910-database-interface.Category'
      - description: apiKey required for non OAuth2 stores only
        in: header
        name: apiKey
        type: string
      - description: store name
        in: header
        name: storeName
        required: true
        type: string
      - description: store localDomain
        in: header
        name: localDomain
        required: true
        type: string
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: OAuth2 client ID only for OAuth2 stores
        in: header
        name: clientId
        type: string
      - description: User ID only for OAuth2 stores
        in: header
        name: userId
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/managers.Response'
      summary: Update a category
      tags:
      - Category
...
swagger: "2.0"